requirements.txt
pillow
numpy
gradio
torch
openai
transformers
dotenv
python-dotenv
datasets
tqdm
opencv-python
modelscope
requests
matplotlib
peft
trl
accelerate
modelscope
qwen_vl_utils
diffusers
torchvision
torchaudio
swanlab
deepspeed